requirements.txt
alembic
autoinject>=1.2.0
azure-core
azure-storage-blob
click
flask
flask-autoinject>=2.0.2
flask-login
flask-wtf
markupsafe
jinja2
pyyaml
requests
sqlalchemy
toml
universalio
zirconium>=1.1.0
zrlog>=0.3.0
azure-identity
azure-storage-file-share
msal
prometheus-client
prometheus-flask-exporter
netCDF4
google-cloud-translate
google-auth
beautifulsoup4
